factor out map util functions from Map component and add further comments detailing what is happening
